requirements.txt
fastapi
uvicorn[standard]
websockets
python-dotenv
numpy
opencv-python-headless
astropy
tensorflow
sgp4
skyfield
Pillow
pandas
